backend/pet_seeker/requirements.txt
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0